app/src/main/java/com/goodylabs/android/interview/ui/characterdetails/CharacterDetailsViewModel.kt
package com.goodylabs.android.interview.ui.characterdetails

import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import com.goodylabs.android.interview.data.models.Character
import com.goodylabs.android.interview.data.repositories.CharacterRepository
import dagger.hilt.android.lifecycle.HiltViewModel
import kotlinx.coroutines.launch
import javax.inject.Inject

@HiltViewModel
class CharacterDetailsViewModel @Inject constructor(
    private val characterRepository: CharacterRepository
) : ViewModel() {

    private val _character = MutableLiveData<Character>()
    val character: LiveData<Character> = _character

    private val _errorText = MutableLiveData<String?>()
    val errorText: LiveData<String?> = _errorText

    fun clearErrorText() {
        _errorText.value = null
    }

    fun refresh(characterId: Int) {
        viewModelScope.launch {
            try {
                _character.value = characterRepository.getCharacterById(characterId)
            } catch (e: Exception) {
                _errorText.value = "Error: $e"
            }
        }
    }
}
